+++ app/src/main/java/com/devmob/alaya/MainContent.kt
@@ -18,7 +18,7 @@ import com.devmob.alaya.domain.model.FeedbackType
 import com.devmob.alaya.domain.model.IconType
 import com.devmob.alaya.domain.model.ItemMenu
 import com.devmob.alaya.ui.components.AppBar
-import com.devmob.alaya.ui.screen.HomeScreen
+import com.devmob.alaya.ui.screen.patient_home.PatientHomeScreen
 import com.devmob.alaya.ui.components.BottomBarNavigation
 import com.devmob.alaya.ui.screen.feedback.FeedbackScreen
 import com.devmob.alaya.ui.screen.login.LoginScreen
@@ -28,7 +28,7 @@ import com.devmob.alaya.ui.screen.ContainmentNetwork.ContainmentNetworkScreen
 import com.devmob.alaya.ui.screen.ContainmentNetwork.ContainmentNetworkViewModel
 import com.devmob.alaya.ui.screen.MenuPatientScreen
 import com.devmob.alaya.ui.screen.MenuProfessionalScreen
-import com.devmob.alaya.ui.screen.PatientHomeScreenViewmodel
+import com.devmob.alaya.ui.screen.patient_home.PatientHomeScreenViewmodel
 import com.devmob.alaya.ui.screen.ProfessionalTreatment.ConfigTreatmentScreen
 import com.devmob.alaya.ui.screen.ProfessionalTreatment.ConfigTreatmentViewModel
 import com.devmob.alaya.ui.screen.TreatmentSummaryScreen.TreatmentSummaryScreen
@@ -107,7 +107,7 @@ fun MainContent(navController: NavHostController) {
                     )
                 }
             ) {
-                HomeScreen(PatientHomeScreenViewmodel(),navController)
+                PatientHomeScreen(PatientHomeScreenViewmodel(),navController)
             }
             composable(NavUtils.ProfessionalRoutes.Home.route,
                 enterTransition = {
